# Add SubArray optimization to sorting