Motivating benchmarks:
```julia
using Random, BenchmarkTools

struct LinArray{T, N} <: AbstractArray{T, N}
    data::Array{T, N}
end
Base.size(a::LinArray) = size(a.data)
Base.@propagate_inbounds Base.getindex(a::LinArray, i...) = getindex(a.data, i...)
Base.@propagate_inbounds Base.setindex!(a::LinArray, v, i...) = setindex!(a.data, v, i...)
Base.IndexStyle(::Type{<:LinArray}) = IndexLinear()
Base.copy(a::LinArray) = LinArray(copy(a.data))

struct CartArray{T, N} <: AbstractArray{T, N}
    data::Array{T, N}
end
Base.size(a::CartArray) = size(a.data)
Base.@propagate_inbounds Base.getindex(a::CartArray, i...) = getindex(a.data, i...)
Base.@propagate_inbounds Base.setindex!(a::CartArray, v, i...) = setindex!(a.data, v, i...)
Base.copy(a::CartArray) = CartArray(copy(a.data))

const master = Base.Sort.DEFAULT_STABLE.next
const pr = Base.Sort.DEFAULT_STABLE

using Random, BenchmarkTools
function bench()
    X = reshape(randperm(40000), 200, 200);
    Y = reshape(randperm(400), 20, 20);
    Z = rand(10000);
    for f in [identity, LinArray, CartArray]
        x, y, z = (f(X), f(Y), f(Z))
        println(typeof(copy(x)))
        @btime sort!(copy($x), dims=1, alg=$master);
        @btime sort!(copy($x), dims=1, alg=$pr);
        println(typeof(copy(y)))
        @btime sort!(copy($y), dims=1, alg=$master);
        @btime sort!(copy($y), dims=1, alg=$pr);
        println(typeof(copy(z)))
        @btime sort!(view(copy($z), 100:1000), alg=$master);
        @btime sort!(view(copy($z), 100:1000), alg=$pr);
    end
end
```

```
julia> bench()
Matrix{Int64}
  222.375 μs (405 allocations: 426.70 KiB)
  190.417 μs (605 allocations: 432.95 KiB)
Matrix{Int64}
  1.208 μs (5 allocations: 3.42 KiB)
  1.229 μs (5 allocations: 3.42 KiB)
Vector{Float64}
  10.708 μs (9 allocations: 87.44 KiB)
  10.250 μs (9 allocations: 87.44 KiB)
LinArray{Int64, 2}
  225.125 μs (406 allocations: 426.72 KiB)
  215.667 μs (406 allocations: 426.72 KiB)
LinArray{Int64, 2}
  1.221 μs (6 allocations: 3.44 KiB)
  1.229 μs (6 allocations: 3.44 KiB)
LinArray{Float64, 1}
  10.750 μs (9 allocations: 87.44 KiB)
  10.458 μs (9 allocations: 87.44 KiB)
CartArray{Int64, 2}
  226.208 μs (406 allocations: 426.72 KiB)
  225.208 μs (406 allocations: 426.72 KiB)
CartArray{Int64, 2}
  1.221 μs (6 allocations: 3.44 KiB)
  1.225 μs (6 allocations: 3.44 KiB)
CartArray{Float64, 1}
  10.791 μs (9 allocations: 87.44 KiB)
  10.875 μs (9 allocations: 87.44 KiB)
```

The reason I'm doing this is I'm trying to make `sort!(::AbstractArray; dims)` faster so that we can delete the `sort(::AbstractArray; dims)` special case for increased codesharing and to make #49394 go away.